Chore: make the server index an actual index
